De-hardcode initialDate and delete Day type
